Moved client.py to src * Refactored imports from clients.py in job.py, sync.py, util.py
