+++ job.py
@@ -2,7 +2,7 @@ import logging
 import sys
 from enum import Enum, auto
 
-from client import Client
+from src.client import Client
 from sync import Sync
 
 log = logging.getLogger(__name__)
